+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.10 -
 
+*   deprecated `_with` contexts [#195](https://github.com/amoffat/sh/issues/195)
 *   `_timeout_signal` allows you to specify your own signal to kill a timed-out process with.  use a constant from the `signal` stdlib module. [#171](https://github.com/amoffat/sh/issues/171)
 *   signal exceptions can now be caught by number or name.  `SignalException_9 == SignalException_SIGKILL`
 *   child processes that timeout via `_timeout` raise `sh.TimeoutException` instead of `sh.SignalExeception_9` [#172](https://github.com/amoffat/sh/issues/172)